+++ src/Presentation/SmartStore.Web/Administration/Models/Settings/CatalogSettingsModel.cs
@@ -45,6 +45,9 @@ namespace SmartStore.Admin.Models.Settings
         [SmartResourceDisplayName("Admin.Configuration.Settings.Catalog.IgnoreDiscounts")]
         public bool IgnoreDiscounts { get; set; }
 
+        [SmartResourceDisplayName("Admin.Configuration.Settings.Catalog.ApplyPercentageDiscountOnTierPrice")]
+        public bool ApplyPercentageDiscountOnTierPrice { get; set; }
+
         [SmartResourceDisplayName("Admin.Configuration.Settings.Catalog.ApplyTierPricePercentageToAttributePriceAdjustments")]
         public bool ApplyTierPricePercentageToAttributePriceAdjustments { get; set; }